Move CloseWithOtherWidgets to Widget
